+++ packages/data-schema/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aws-amplify/data-schema
 
+## 1.8.1
+
+### Patch Changes
+
+- 29f13a4: Expose name, metadata, createdAt, updatedAt in Covnersation Type
+
+## 1.8.0
+
+### Minor Changes
+
+- 6eb40fc: Add conversation delete api, re-organized AiAction enums to align with Amplify-JS"
+
 ## 1.7.0
 
 ### Minor Changes